Made affiliations and impact_areas have a single string field
This field contains the name of the affiliation or impact area.  Before,
each affiliation or impact area was a single column in the database.
This meant changing the schema whenever a new affiliation or impact area
was added.  It also meant that the schema was very MUSC specific.  This
changes addresses both of these issues.