Load schema before calling FactoryGirl.find_definitions
A change was made to the services factory which caused
FactoryGirl.find_definitions to fail with an error about unable to find
column pricing_maps.  Fixed this by loading the schema before calling
find_definitions.  I do not know why this is necessary now but was not
necessary before now.

I am using an in-memory sqlite database.  I do not know how this change
will affect testing with a mysql database.