Fixed test so it will pass and added a new test
Test was checking to make sure subject count is not set if it is already
set (on a per patient per visit line item).

A new condition was added that DOES set subject count if it is higher
than the arm's subject count.

We now check both possibilities: the vg's visit count is higher than the
arm's subject count, and the vg's visit count is lower than the arm's
subject count.