Ok, I think I fixed the test for real this time
I think the problem was in the before(:each) block; it wasn't waiting
for the click to complete before starting the test.  Added some other
calls to wait_for_javascript_to_finish just to be on the safe side.

Also removed the call to sleep(2); the tests seem to pass just fine
without it.