wth - Admin edit sending SSR back to Draft bug
When altering figures in the service calendar tables, the statuses were
being changed back to draft status. This is proper functionality for
proper, but not for Dashboard. I added a parameter to the update actions
for Visits and Line Items Visits that checks if the update is happening
in Dashboard. [#136778033]