wth - Step 4: Added Protocol Notes not showing up
On SPARCRequest, the added protocol notes on Step 3 are not showing up
on Step 4 (summary) page, although the data is stored correctly and the
notes show up on SPARCDashboard.

This was caused by the Notable type and ID being set to 'ServiceRequest'
and the ID of the ServiceRequest, when it should've been to the
Protocol. [#139499703]