wth - (SPARCRequest & SPARCDashboard) Tooltips: Step 1 Study Information
Added tooltips to the SPARCRequest Step 1 Protocol
Information Page and Dashboard edit page. Some of the bootstrap selects
were acting weird with normal markup so I wrapped them in anchor tags,
seems to work fine. Open to suggestions on the JS code in this
PR, does the job but maybe there is a better way if anyone has ideas. [#140086745]