wth - SPARCCatalog Pricing Map Rates (Section): Delete Override Price Revamp
In SPARCCatalog, when inside the pricing map section of a service and
trying to delete an existing override price, and click "Save", there was
no error message, however, the change was not saved because it was a NaN
(Not a Number).

I added jQuery code to check if the input is blank and if so, replace it
with the calculated rate. [#127112423]