wth - (SPARCDashboard) Saved Filters Bug
Since the latest production (v2.0.0), the previous saved filters on
SPARCDashboard are no longer working.

This was caused by an error in the search_query scope. It was expecting
a Hash but was saved as a string. I attempted first to serialize the
attribute as a Hash while converting the data-type to Text, but that was
giving me issues. I ended up just using the #eval method to convert to
hash on ProtocolFilter#href. Also wrote a rake task to fix historical
filters and wrote feature spec where a user interacts with saved
searches. [#142261025]