wth - (SPARCRequest & SPARCDashboard) Switching PI Bug
On SPARCRequest and SPARCDashboard, when switching PIs by adding a new
primary PI on the protocol, it is defaulting the previous PI to
"Request/Approve Services" right, which doesn't exist any more.

I updated the two classes in /lib that handle this update and set the
project rights to 'approve'. [#145688447]

Pivotal Story - (https://www.pivotaltracker.com/story/show/145688447)

Steps to Test:
1. Find existing Protocol
2. Add a new Primary PI, which will replace the old one.
3. Proxy rights should remain as they were.

updated specs