wth - (SPARCDashboard) Add "Edit" Epic Question / CofC Answers for Authorized Users
Background: On SPARCDashboard, when a user clicks "Edit Study Information" on the Study Summary tab, currently, the existing answers to the Epic questions (when use_epic is TRUE) or CofC questions (when use_epic is FALSE) are view-only. This was done to enable the admin users edit other fields study information and save without having to answer those questions. However, this is push the users to have to use "Add/Modify Request" button to modify the answers from SPARCRequest and waste time if that's the only part they want to change.

1). Create an "Edit" button next to the "Publish Study in Epic" question (when epic is turned on) and next to the first CofC question (when epic is turned off),
and 2) allow the users to edit the existing answers when clicked upon.
3). The new "Edit" button should show up for everyone, study team or admins, who has the rights to the chosen protocol.

[#148375589]

Story - https://www.pivotaltracker.com/story/show/148375589