wth & jm - Change datatype DelayedJob Handler from Text to LongText
Some exceptions received by Delayed Job can be quite lengthy. If you are using MySQL, the handler and last_error fields may not be long enough. Change their datatype to longtext to avoid this issue. If you are using PostgreSQL, this will not be a problem.

https://www.sitepoint.com/delayed-jobs-best-practices/
https://stackoverflow.com/questions/4443477/rails-3-migration-with-longtext

[#148490549]

Story - https://www.pivotaltracker.com/story/show/148490549