wth - (SPARCCatalog) Form Functionality Bugs
Inside of the form builder, the description(s) (when added) are not displaying to the user.
Secondly, the sentence structure does not allow caps to be displayed to the user, although letters are capitalized in the form builder (i.e. FOA, TIN, etc).
Lastly, question numbering does not display on the questionnaire side when being entered on the form builder. Allow numbering to display so questions can be numbered.

[#149884606]

Story - https://www.pivotaltracker.com/story/show/149884606