wth - (SPARCDashboard) Tie Editable Fields with RMID Configuration
When RMID configuration is turned on (RESEARCH_MASTER_ENABLED == false), and a protocol has a RMID previously, there is currently no way to edit the Short Title or the Long Title.
Please tie the editing functions with the RMID configuration, so that the fields become editable when the configuration is turned off.

[#150408490]

Story - https://www.pivotaltracker.com/story/show/150408490