wth - (SPARCRequest & SPARCDashboard) Update/Pull Pro# for Validated Protocols
Background: Once a protocol has a RMID associated, and that RMID is validated (because its eIRB record is one of the validated states), the short title and long title has now being updated via a cronjob, and the "Updated to corresponding Research Master ID record...." label next to the fields to indicate.

Please add the functionality of updating Pro# according to the eIRB Pro# associated with a validated eIRB record via RMID of the protocol, and display the validation visual cue.

[#151209478]

Story - https://www.pivotaltracker.com/story/show/151209478

added cue