wth - (SPARCRequest & SPARCDashboard) Tie Display of Indirect Cost Rate with Configuration
Background: The indirect cost calculations have been tied with the configuration, so that when the configuration is turned off, the rows of indirect calculations don't show up. However, in the Protocol information section, the "Indirect Cost Rate" field (not-required) is still showing up for users to fill in, which is confusing and seems.

Please tie the display of the Indirect Cost Rate with configuration (use_indirect_cost), too; So that when use_indirect_cost: false, the field doesn't display on SPARCRequest Step 2 and SPARCDashboard Protocol Information section. See attached screenshot for more details.

[#151174598]

Story - https://www.pivotaltracker.com/story/show/151174598