wth - (SPARCRequest & SPARCDashboard) Epic User Update New Feature
Background: After a protocol is pushed to Epic, and new user comes into the study team, currently it is a manual process to update both in SPARC and Epic with the new user.

When a new user who has Epic access (project_roles.epic_access = 1) is being "Added" to a protocol that has been pushed to epic before (protocols.selected_for_epic = 1; and protocols. last_epic_push_time is NOT NULL), a SOAP message will pushed to Epic with only the protocol information and the identity information without the timeline (CYCLE).

1). Please queue the authorized user change (on protocols that have already been pushed to Epic) and send it out daily to Epic.

2). Please display the protocols that have the authorized user change on a tab on the "Epic Queue" page.

[#145759415]

Story - https://www.pivotaltracker.com/story/show/145759415