wth - (SPARC/RMID API) Handling Deleted RMID Record
Background: When a RMID is used by a SPARC protocol, that SPARC protocol is associated with the RMID, with short title and long title pulled from RMID record, and pro# pulled from the associated eIRB record (if it exists). However, when the RMID record is deleted, currently it's failing the API, because the long title, short title, pro# on the SPARC protocol will be updated to blank. Ideally, the user should update the RMID in SPARC, but this scenario should be taken into consideration.

Please change the API rules, so that when the RMID a SPARC protocol uses doesn't exist any more, remove the RMID in SPARC database when API refreshes, but keep the long title, short title, and pro#.

Acceptance criteria:
1). When a user deletes a RMID that already has associated SPARC record, no fatal error happens;
2). The current long title, short title, and pro# should be kept;
3). This method should be tied with RMID configuration for turning on/off.

[#151617565]

Story - https://www.pivotaltracker.com/story/show/151617565