wth - (SPARCRequest & SPARCDashboard) Epic User Update/Deletion Feature
With a previous story (https://www.pivotaltracker.com/story/show/145759415), when there is a user added with epic rights on an existing protocol (that has been pushed to Epic), or epic rights have been added to an existing user, SPARC is now automatically queuing the protocol and sending only protocol information over to Epic daily.

However, 1). when a user with Epic rights is removed from existing Epic protocols, or 2). when a user with Epic rights is changed to no Epic rights, it is not triggering the protocol info to update.

Please add the 2 criteria into the method, so that we are covered with any user changes related to Epic rights being updated.

[#152091104]

Story - https://www.pivotaltracker.com/story/show/152091104