wth - (SPARCDashboard) Validate Calendar when Sending Request to SPARCFulfillment
Background: For service providers who are using SPARCFulfillment module, currently when the user clicks the "Send to Fulfillment" button on SPARCDashboard to push the request to SPARCFulfillment, it is not taking into consideration whether the calendar is intact. If the protocol has a broken calendar (i.e. missing visit days, out of sequence visits, etc), it is currently causing the SSR to not show in SPARCFulfillment without any explanation or error message.

Acceptance criteria:
1). Add validation on the "Send to Fulfillment" button to check on the calendar integrity.
2). If the protocol has a broken calendar, a modal with error message "This request has failed to be sent to SPARCFulfillment because of failed validation. Please make sure the service calendar is intact before trying again."
3). The button should stay "Send to Fulfillment" if the send has failed, so that the user could try again.

[#151879348]

Story - https://www.pivotaltracker.com/story/show/151879348