wth - (SPARCDashboard) Epic Queue Past Tab "Type" Bug
The "Past" column on Epic Queue view on SPARCDashboard is showing the wrong "Type" for some of the pushes. For example, the screenshot attached is showing a general user who did the admin push on 4/4/2018. When I invested into the audit trail, it shows that this study team user deleted 3 project roles with epic access on 4/4/2018. So it seems that it should have been an "Protocol Update" as type.

Please look into this bug, and double-check all the types are tagged correctly, and fix the data that has been affected.

[#156541424]

Story - https://www.pivotaltracker.com/story/show/156541424