wth - (SPARCForm) Response View Filters
I fixed saved filter searches. The issue was that we were saving
start_date and end_date as timestamps, when it is easier for filterrific
to process strings. Rails auto saves dates as 'yyyy-mm-dd', so I wrote a
method that updates it to correct format.

Any feedback is welcome.

[#156401749]

Story - https://www.pivotaltracker.com/story/show/156401749