Adds cost analysis PDF report.
This is for story https://www.pivotaltracker.com/story/show/161627471

This changeset adds a new 'Cost Analysis Report' button to the
SPARC dashboard.

The report tries to reuse as much of the existing logic as it
could from existing dashboard views but then produces a report
that breaks down line item visit counts and charges by arm.

The visit tables try to compute the best fit size (column width
and number of columns) to use, as well as try to prevent spanning
pages if possible.